feat(py/algod_client): DefaultHttpClient
